refactor(#176): move config migrations to a file
