+++ src/modules/ConfigStorage/migrations.ts
@@ -1,0 +1,91 @@
+import browser from 'webextension-polyfill';
+
+import {
+	configureMigration,
+	MigrationObject,
+	MigrationTask,
+} from '../../migrations/migrations';
+
+const migrations: MigrationObject[] = [
+	{
+		version: 1,
+		async migrate() {
+			const storageKey = 'config.Main';
+			const storageDataRaw = localStorage.getItem(storageKey);
+
+			// Skip
+			if (storageDataRaw === null) return;
+
+			const storageNameV2 = 'appConfig';
+
+			// Import valid data
+			const storageData = JSON.parse(storageDataRaw);
+			if (typeof storageData === 'object') {
+				// Merge actual data with legacy
+				let { [storageNameV2]: actualData } = await browser.storage.local.get(
+					storageNameV2,
+				);
+				if (typeof actualData !== 'object') {
+					actualData = {};
+				}
+
+				const mergedData = { ...actualData, ...storageData };
+
+				// Write data
+				browser.storage.local.set({
+					[storageNameV2]: mergedData,
+				});
+			}
+
+			// Delete old data
+			localStorage.removeItem(storageKey);
+		},
+	},
+	{
+		version: 2,
+		async migrate() {
+			const storageNameV2 = 'appConfig';
+
+			// Merge actual data with old
+			let { [storageNameV2]: actualData } = await browser.storage.local.get(
+				storageNameV2,
+			);
+			if (typeof actualData !== 'object') {
+				actualData = {};
+			}
+
+			const contentscriptPropData =
+				actualData?.contentscript?.selectTranslator || {};
+			const quickTranslate = actualData?.selectTranslator?.quickTranslate;
+
+			const newData = actualData;
+			delete newData.contentscript;
+
+			if (newData.selectTranslator) {
+				delete newData.selectTranslator.quickTranslate;
+			}
+
+			// Write data
+			browser.storage.local.set({
+				[storageNameV2]: {
+					...newData,
+					selectTranslator: {
+						...newData?.selectTranslator,
+						...contentscriptPropData,
+						mode: quickTranslate
+							? 'quickTranslate'
+							: newData?.selectTranslator?.mode,
+					},
+				},
+			});
+		},
+	},
+];
+
+export const ConfigStorageMigration: MigrationTask = {
+	version: 3,
+	async migrate(prevVersion) {
+		const migrate = configureMigration(migrations);
+		await migrate({ fromVersion: prevVersion, toVersion: 3 });
+	},
+};
